% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{More interesting pop-ups for the interactive map used with the eq_map() function}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{Cleaned NOAA data.frame.}
}
\value{
HTML labels to annotate leaflet map.
}
\description{
More interesting pop-ups for the interactive map used with the eq_map() function
}
\examples{
\dontrun{
readr::read_delim("eq.txt", delim = "\t") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}

}
