% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleaning NOAA locations}
\usage{
eq_location_clean(loc)
}
\arguments{
\item{loc}{Raw data.frame with data from NOAA website.}
}
\value{
The data.frame with cleaned LOCATION_NAME.
}
\description{
Cleaning NOAA locations
}
\examples{
\dontrun{
library(readr)
df = readr::read_delim("eq.txt", delim = "\t")
clean_df = eq_location_clean(df)
}

}
