% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_map}
\alias{eq_map}
\title{Marking and annotating earthquake epicenters on the map.}
\usage{
eq_map(df, annot_col = "DATE")
}
\arguments{
\item{df}{Cleaned NOAA data.frame.}

\item{annot_col}{Column used for annotation}
}
\value{
Leaflet map of the earthquakes epicenters with annotations.
}
\description{
Marking and annotating earthquake epicenters on the map.
}
\examples{
\dontrun{
readr::read_delim("eq.txt", delim = "\t") \%>\%
eq_clean_data() \%>\% 
eq_location_clean \%>\% 
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")
}

}
