% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Timeline for earthquakes}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetics created by aes.}

\item{data}{Cleaned NOAA data.frame}

\item{stat}{A string of statistical transformation.}

\item{position}{Position adjustment.}

\item{na.rm}{A logical for dealing with missing values.}

\item{show.legend}{A logical for showing legend or not.}

\item{inherit.aes}{A logical of whether or not override default aesthetics.}

\item{...}{Other arguments passed to layer}
}
\description{
Timeline for earthquakes
}
\details{
A timeline is a visualization of earthquakes ordered by date of occurrence, point's color corresponds to the number of deaths
and the size represents the magnitude of earthquake.
}
\examples{
\dontrun{
clean_df \%>\%
dplyr::filter(COUNTRY == c("MEXICO","USA") & lubridate::year(DATE) >= 2010) \%>\%
ggplot(aes(x = DATE,
           y = COUNTRY,
           color = TOTAL_DEATHS,
           size = EQ_PRIMARY)) +
geom_timeline(alpha=.7) +
theme(legend.position = "bottom", legend.box = "horizontal", plot.title = element_text(hjust=0.5)) +
ggtitle("Earthquakes timeline") +
labs(size = "Richter scale value", color = "# deaths")
}
}
