% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Timeline label}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetics created by aes.}

\item{data}{Cleaned NOAA data.frame}

\item{stat}{A string of statistical transformation.}

\item{position}{Position adjustment.}

\item{na.rm}{A logical for dealing with missing values.}

\item{show.legend}{A logical for showing legend or not.}

\item{inherit.aes}{A logical of whether or not override default aesthetics.}

\item{...}{Other arguments passed to layer.}
}
\description{
Timeline label
}
\details{
Adds labels for the locations of individual earthquaks. Used with geom_timeline.
}
\examples{
\dontrun{
clean_df \%>\%
dplyr::filter(COUNTRY == c("MEXICO","USA") & lubridate::year(DATE) >= 2000) \%>\%
ggplot(aes(x = DATE,
           y = COUNTRY,
           color = TOTAL_DEATHS,
           size = EQ_PRIMARY)) +
geom_timeline(alpha=.7) +
geom_timeline_label(aes(label = LOCATION_NAME), n_max=4) +
theme(legend.position="bottom", legend.box="horizontal", plot.title=element_text(hjust=0.5)) +
ggtitle("Earthquakes timeline") +
labs(size = "Richter scale value", color = "# deaths")
}
}
