% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleaning NOAA dates, geographic coordinates and location name}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{Raw data.frame with data from NOAA website.}
}
\value{
Cleaned data.frame with formatted date, longitude/latitude columns and location name.
}
\description{
Cleaning NOAA dates, geographic coordinates and location name
}
\examples{
\dontrun{
df = readr::read_delim("eq.txt", delim = "\t")
clean_df = eq_clean_data(df)
}

}
